+++ src/main/kotlin/uk/gov/justice/digital/hmpps/incentivesapi/dto/prisonapi/PrisonAPI.kt
@@ -1,8 +1,6 @@
 package uk.gov.justice.digital.hmpps.incentivesapi.dto.prisonapi
 
 import org.springframework.format.annotation.DateTimeFormat
-import uk.gov.justice.digital.hmpps.incentivesapi.service.PrisonerInfoForNextReviewDate
-import java.time.LocalDate
 import java.time.LocalDateTime
 
 data class Prison(
@@ -12,19 +10,6 @@ data class Prison(
   val agencyType: String,
   val active: Boolean,
 )
-
-data class PrisonerAlert(
-  val alertType: String,
-  val alertCode: String,
-  val active: Boolean,
-  val expired: Boolean,
-) {
-  companion object {
-    const val ACCT_ALERT_CODE = "HA"
-  }
-
-  val isOpenAcct = alertCode == ACCT_ALERT_CODE && active && !expired
-}
 
 data class PrisonerCaseNoteByTypeSubType(
   val bookingId: Long,
